pyproject.toml
[build-system]
requires = ["setuptools>=65.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "recommendation-system"
version = "1.0.0"
description = "Production-grade recommendation system using MovieLens 1M"
requires-python = ">=3.11"
authors = [{name = "MLOps Team"}]

dependencies = [
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "scikit-learn>=1.3.0",
    "tensorflow>=2.14.0",
    "torch>=2.0.0",
    "pytorch-lightning>=2.0.0",
    "feast>=0.35.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "hydra-core>=1.3.0",
    "omegaconf>=2.3.0",
    "optuna>=3.0.0",
    "mlflow>=2.0.0",
    "boto3>=1.26.0",
    "redis>=5.0.0",
    "pinecone-client>=3.0.0",
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "streamlit>=1.28.0",
    "python-dotenv>=1.0.0",
    "structlog>=23.0.0",
    "python-multipart>=0.0.6",
    "aiohttp>=3.9.0",
    "httpx>=0.25.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.5.0",
    "isort>=5.12.0",
]

[tool.black]
line-length = 100
target-version = ['py311']

[tool.ruff]
line-length = 100
target-version = "py311"
select = ["E", "F", "W"]
ignore = ["E501"]

[tool.isort]
profile = "black"
line_length = 100
